pyproject.toml
[tool.poetry]
name = "ham_tools"
version = "0.1.0"
description = "Various amateur radio tools"
authors = ["Nick Pegg <nick@nickpegg.com>"]
license = "MIT"

[tool.poetry.scripts]
ft991a = 'ham_tools.cli.ft991a:main'
cat_shell = 'ham_tools.cli.cat_shell:main'

[tool.poetry.dependencies]
python = ">=3.8.0,<4"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
black = "*"
mypy = "*"
pytest = "*"
isort = "*"
ipython = "*"

[tool.isort]
profile = "black"

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = "serial.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = [
  "integration: Integration tests which require a real radio"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
